feat: add caps lock functionality for keys
