+++ virtual-keyboard/index.html
@@ -15,7 +15,7 @@
   <main class="page">
     <article class="keyboard">
       <form class="keyboard__form">
-        <textarea  name="" id="" cols="104" rows="7" class="keyboard__textarea" autofocus></textarea>
+        <textarea  name="" id="" cols="104" rows="7" class="keyboard__textarea"></textarea>
       </form>
       <div class="keys">
         <div class="keys__line">
@@ -96,6 +96,6 @@
       </div>
     </article>
   </main>
-  <script src="js/print-events.js"></script>
+  <script src="index.js"></script>
 </body>
 </html>
